[minor] Fix bug in RPROMPT not showing parent's name correctly
We extract the parent's command name by using ps -p -ocomm= $PPID
The = is required after comm because it omits the header row.
This way, the command expansion shows only the command's name, which
is the intended behaviour.

As of why no text at all was displayed (instead of, for example, COMM login)
is still a mistery to me; but at least it works now.